refactor: tobe 패키지 제거
