+++ src/main/java/kitchenpos/menus/application/MenuGroupService.java
@@ -1,9 +1,9 @@
 package kitchenpos.menus.application;
 
 import kitchenpos.menus.model.MenuGroupModel;
-import kitchenpos.menus.tobe.domain.MenuGroup;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
-import kitchenpos.menus.tobe.domain.MenuGroupName;
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
+import kitchenpos.menus.domain.MenuGroupName;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
