+++ src/main/java/kitchenpos/menus/domain/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.domain;
 
 import javax.persistence.*;
 import java.util.UUID;
