+++ src/main/java/kitchenpos/menus/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.domain;
 
 import java.util.List;
 import java.util.Optional;
