+++ src/main/java/kitchenpos/menus/domain/MenuProductQuantity.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.domain;
 
 import javax.persistence.Embeddable;
 import java.math.BigDecimal;
