+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -1,8 +1,8 @@
 package kitchenpos.products.application;
 
-import kitchenpos.menus.tobe.domain.Menu;
+import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuProduct;
-import kitchenpos.menus.tobe.domain.MenuRepository;
+import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.products.model.ProductModel;
 import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
