+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -3,10 +3,9 @@ package kitchenpos.menus.application;
 import kitchenpos.menus.model.MenuModel;
 import kitchenpos.menus.model.MenuProductRequest;
 import kitchenpos.menus.model.MenuRequest;
-import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
-import kitchenpos.menus.domain.MenuProduct;
-import kitchenpos.menus.tobe.domain.MenuRepository;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuGroupRepository;
+import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.products.application.FakePurgomalumClient;
 import kitchenpos.products.application.InMemoryProductRepository;
 import kitchenpos.products.domain.ProductRepository;
